integrate mpi and distribution components
